fix: refactor skip_on decorator and update Python version requirement in pyproject.toml
