+++ argilla/pyproject.toml
@@ -7,7 +7,7 @@ authors = [
 maintainers = [
     { name = "Extralit Labs", email = "extralit.contact@gmail.com" },
 ]
-requires-python = ">=3.9.0, !=3.9.1, <3.14"
+requires-python = ">=3.9.2, <3.14"
 readme = "README.md"
 license = { text = "Apache 2.0" }
